+++ core/services/fluxmonitor/flux_monitor_simulated_blockchain_test.go
@@ -569,3 +569,52 @@ func TestFluxMonitor_HibernationMode(t *testing.T) {
 	case <-time.After(5 * time.Second):
 	}
 }
+
+func TestFluxMonitor_InvalidSubmission(t *testing.T) {
+	// Comments starting with "-" describe the steps this test executes.
+	key := cltest.MustGenerateRandomKey(t)
+	fa := setupFluxAggregatorUniverse(t, key)
+	oracleList := []common.Address{fa.neil.From, fa.ned.From, fa.nallory.From}
+	_, err := fa.aggregatorContract.ChangeOracles(fa.sergey, emptyList, oracleList, oracleList, 1, 3, 2)
+	assert.NoError(t, err, "failed to add oracles to aggregator")
+	fa.backend.Commit()
+
+	// Set up chainlink app
+	config, cfgCleanup := cltest.NewConfig(t)
+	config.Config.Set("DEFAULT_HTTP_TIMEOUT", "100ms")
+	config.Config.Set("TRIGGER_FALLBACK_DB_POLL_INTERVAL", "1s")
+	config.Config.Set("MIN_OUTGOING_CONFIRMATIONS", "2")
+	config.Config.Set("MIN_OUTGOING_CONFIRMATIONS", "2")
+	config.Config.Set("ETH_HEAD_TRACKER_MAX_BUFFER_SIZE", "100")
+	defer cfgCleanup()
+	app, cleanup := cltest.NewApplicationWithConfigAndKeyOnSimulatedBlockchain(t, config, fa.backend, key)
+	defer cleanup()
+	require.NoError(t, app.StartAndConnect())
+
+	// Report a price that is above the maximum allowed value,
+	// causing it to revert.
+	reportPrice := int64(1000000000000)
+	priceResponse := func() string {
+		return fmt.Sprintf(`{"data":{"result": %d}}`, atomic.LoadInt64(&reportPrice))
+	}
+	mockServer := cltest.NewHTTPMockServerWithAlterableResponse(t, priceResponse)
+	defer mockServer.Close()
+	buffer := cltest.MustReadFile(t, "../../internal/testdata/flux_monitor_job.json")
+	var job models.JobSpec
+	require.NoError(t, json.Unmarshal(buffer, &job))
+	initr := &job.Initiators[0]
+	initr.InitiatorParams.Feeds = cltest.JSONFromString(t, fmt.Sprintf(`["%s"]`, mockServer.URL))
+	initr.InitiatorParams.PollTimer.Period = models.MustMakeDuration(100 * time.Millisecond)
+	initr.InitiatorParams.Address = fa.aggregatorContractAddress
+
+	j := cltest.CreateJobSpecViaWeb(t, app, job)
+	go func() {
+		for {
+			fa.backend.Commit()
+			time.Sleep(500 * time.Millisecond)
+		}
+	}()
+	// We should see a spec error because the value is too large to submit on-chain.
+	jse := cltest.WaitForSpecError(t, app.Store, *j.ID, 1)
+	assert.Contains(t, jse[0].Description, "Polled value is outside acceptable range")
+}